# docs/swagger.json
{
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is an example server using Sochain API at backend",
        "title": "Sochain API Explorer",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8081",
    "basePath": "/api/v1",
    "paths": {
        "/block/{network}/{blockNumberOrHash}": {
            "get": {
                "description": "Show block by network \u0026 number/hash",
                "tags": [
                    "blocks"
                ],
                "summary": "Show block",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Network",
                        "name": "network",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "block number or hash",
                        "name": "blockNumberOrHash",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/service.DesiredBlockResponseData"
                        }
                    }
                }
            }
        },
        "/tx/{network}/{hash}": {
            "get": {
                "description": "Show transaction by network \u0026 hash",
                "tags": [
                    "transactions"
                ],
                "summary": "Show transaction",
                "parameters": [
                    {
                        "type": "string",
                        "description": "network",
                        "name": "network",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "transaction hash",
                        "name": "hash",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/service.DesiredTxResponseData"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "service.DesiredBlockResponseData": {
            "type": "object",
            "properties": {
                "block_no": {
                    "type": "integer"
                },
                "network": {
                    "type": "string"
                },
                "next_blockhash": {
                    "type": "string"
                },
                "previous_blockhash": {
                    "type": "string"
                },
                "size": {
                    "type": "integer"
                },
                "time": {
                    "description": "actual type is `int`, but we need to display this time in string format in API response",
                    "type": "string"
                },
                "txs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/service.DesiredTxResponseData"
                    }
                }
            }
        },
        "service.DesiredTxResponseData": {
            "type": "object",
            "properties": {
                "fee": {
                    "type": "string"
                },
                "sent_value": {
                    "type": "string"
                },
                "time": {
                    "type": "string"
                },
                "txid": {
                    "type": "string"
                }
            }
        }
    }
}